--- baekjoon/큰솔/2557. Hello World/README.md
@@ -1,31 +1,0 @@
-# [Bronze V] Hello World - 2557 
-
-[문제 링크](https://www.acmicpc.net/problem/2557) 
-
-### 성능 요약
-
-메모리: 13968 KB, 시간: 116 ms
-
-### 분류
-
-구현
-
-### 제출 일자
-
-2022년 9월 21일 18:08:45
-
-### 문제 설명
-
-<p>
-	Hello World!를 출력하시오.</p>
-
-### 입력 
-
- <p>
-	없음</p>
-
-### 출력 
-
- <p>
-	Hello World!를 출력하시오.</p>
-
